+++ src/themes/base/vuetify.options.js
@@ -1,3 +1,6 @@
+import hourglassBottomBlack from "@assets/icons/hourglass-bottom-black.vue";
+import hourglassDisabled from "@assets/icons/hourglass-disabled.vue";
+
 export default {
 	theme: {
 		themes: {
@@ -11,9 +14,13 @@ export default {
 	},
 	icons: {
 		iconfont: "fa4",
+		values: {
+			hourglassBottomBlack: {
+				component: hourglassBottomBlack,
+			},
+			hourglassDisabled: {
+				component: hourglassDisabled,
+			},
+		},
 	},
-	// possible in vuetify 3
-	/* global: {
-		ripple: false,
-	}, */
 };